Increased the functionality of create_element
Previously create_element would generate an element based on a developer
inputted tag name and option text. Now it takes in tag name, however
instead of text it can take in a whole node, and can also take in an ID
and a list of classes to assign to the newly generated node.